requirements.txt
black
click
Flask
Flask-Migrate
flask-shell-ipython
Flask-SQLAlchemy
gunicorn
ipython
omdb
psycopg2
python-dotenv
requests
selenium
pre-commit
